Move using-spring-web tests to the same package as sources
